Aplicação do Single Responsability Principle (SOLID) #52
